Features - Added automatic key pair authentification (fixed it and added it to client setup) - More robust client setup code
Bug fix
- Fixed server ip address not updating
- Fixed requirements not updated in client
- Fixed "exit" command not working
- Fixed few prints not being computer.log
- Fixed print of input field

Refactored a class
