+++ src/client/requirements_client.txt
@@ -1,2 +1,3 @@
 pywin32
 psutil~=5.9.4
+requests==2.29.0
